fix: Lost reference after connection when scan is started
If starting a scan simultaneously with a connection, the peripheral instance may be removed during the connection process. Depending on the timing, the peripheral can be either missing or replaced with a new instance when trying to use the device. If the instance is replaced, it has a null GATT instance. Thus, any calls using the GATT will fail. This fix guards the peripherals list during the connection and sets the "connecting" state of the peripheral in the same thread the connec function was called. Previously the state was set "late" in the main handler. This should guarantee that the peripheral instance is not removed during the connection process (after calling connect).